[Refactor] 멤버탈퇴 위치변경
냉장고 멤버 탈퇴 refrigerator -> member 로 위치변경
