+++ src/main/kotlin/team/b2/bingojango/domain/member/controller/MemberController.kt
@@ -6,6 +6,7 @@ import org.springframework.http.ResponseEntity
 import org.springframework.security.access.prepost.PreAuthorize
 import org.springframework.security.core.annotation.AuthenticationPrincipal
 import org.springframework.web.bind.annotation.*
+import team.b2.bingojango.domain.member.dto.MemberResponse
 import team.b2.bingojango.domain.member.service.MemberService
 import team.b2.bingojango.global.security.util.UserPrincipal
 
@@ -14,6 +15,18 @@ import team.b2.bingojango.global.security.util.UserPrincipal
 class MemberController(
         private val memberService: MemberService
 ) {
+    @Operation(summary = "냉장고 참여 멤버 조회")
+    @PreAuthorize("isAuthenticated()")
+    @GetMapping()
+    fun getMembers(
+            @AuthenticationPrincipal userPrincipal: UserPrincipal,
+            @PathVariable refrigeratorId: Long
+    ): ResponseEntity<List<MemberResponse>>{
+        return ResponseEntity
+                .status(HttpStatus.OK)
+                .body(memberService.getMembers(userPrincipal, refrigeratorId))
+    }
+
     @Operation(summary = "STAFF 권한 부여")
     @PreAuthorize("isAuthenticated()")
     @PatchMapping("/{memberId}")
