+++ src/main/kotlin/team/b2/bingojango/domain/refrigerator/controller/RefrigeratorController.kt
@@ -7,7 +7,6 @@ import org.springframework.http.ResponseEntity
 import org.springframework.security.access.prepost.PreAuthorize
 import org.springframework.security.core.annotation.AuthenticationPrincipal
 import org.springframework.web.bind.annotation.*
-import team.b2.bingojango.domain.member.dto.MemberResponse
 import team.b2.bingojango.domain.refrigerator.dto.request.AddRefrigeratorRequest
 import team.b2.bingojango.domain.refrigerator.dto.request.JoinByInvitationCodeRequest
 import team.b2.bingojango.domain.refrigerator.dto.request.JoinByPasswordRequest
@@ -66,17 +65,6 @@ class RefrigeratorController(
         return ResponseEntity
             .status(HttpStatus.OK)
             .body(refrigeratorService.joinRefrigeratorByInvitationCode(userPrincipal, request))
-    }
-
-    @Operation(summary = "냉장고 참여 멤버 조회")
-    @PreAuthorize("isAuthenticated()")
-    @GetMapping("/{refrigeratorId}")
-    fun getMembers(
-            @PathVariable refrigeratorId: Long
-    ): ResponseEntity<List<MemberResponse>>{
-        return ResponseEntity
-                .status(HttpStatus.OK)
-                .body(refrigeratorService.getMembers(refrigeratorId))
     }
 
 }
